fixed crash without bloodmagic but with thaumcrafft caused by class loading
